Splitting DataParser into multiple files
